[BUGFIX] Fix handling of millisecond values (#48)
When trying an SSML Break time of 0.5 under Adhearsion, FreeSWITCH
treats this as 0 seconds. This makes some sense as the spec does not
say that you can pass fractional seconds, only whole seconds or
milliseconds with the appropriate suffix.

This change adds the appropriate suffix when rendering XML and
converts back to seconds again when reading XML.